chore(test): improve setup and aliases
