+++ package.json
@@ -39,7 +39,7 @@
 		"typescript": "^5.8.3",
 		"vite": "^6.3.5",
 		"vue-eslint-parser": "^10.2.0",
-		"vitest": "^1.5.0",
-		"@vitest/coverage-v8": "^1.5.0"
+                "vitest": "^3.2.0",
+                "@vitest/coverage-v8": "^3.2.0"
 	}
 }
