+++ vitest.config.ts
@@ -1,24 +1,10 @@
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { defineConfig } from 'vitest/config';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+import aliases from './aliases';
 
 export default defineConfig({
-	resolve: {
-		alias: {
-			'@': path.resolve(__dirname, './src'),
-			'@css': path.resolve(__dirname, './src/css'),
-			'@pages': path.resolve(__dirname, './src/pages'),
-			'@fonts': path.resolve(__dirname, './src/fonts'),
-			'@images': path.resolve(__dirname, './src/images'),
-			'@public': path.resolve(__dirname, './src/public'),
-			'@partials': path.resolve(__dirname, './src/partials'),
-			'@stores': path.resolve(__dirname, './src/stores'),
-			'@api': path.resolve(__dirname, './src/stores/api'),
-		},
-	},
+        resolve: {
+                alias: aliases,
+        },
 	test: {
 		environment: 'jsdom',
 		coverage: {
